Move Modal and Playground components to common/
