+++ plugins/react/frontend/components/common/Modal/__test__/index.js
