+++ plugins/react/frontend/components/common/Modal/index.js
