+++ plugins/react/frontend/components/common/Modal/styles.css
